Move old code to a separate file
